Move `fp::vector` internals around
